Remove references to Prawn constant within PDF::Core namespace
