+++ lib/prawn/font.rb
@@ -223,7 +223,7 @@ module Prawn
 
         # Build up an Arranger with the entire string on one line, finalize it,
         # and find its width.
-        arranger = PDF::Core::Text::Formatted::Arranger.new(self, options)
+        arranger = Prawn::Text::Formatted::Arranger.new(self, options)
         arranger.consumed = self.text_formatter.format(string, *p)
         arranger.finalize_line
 
@@ -238,7 +238,7 @@ module Prawn
     def width_of_inline_formatted_string(string, options={})
       # Build up an Arranger with the entire string on one line, finalize it,
       # and find its width.
-      arranger = PDF::Core::Text::Formatted::Arranger.new(self, options)
+      arranger = Prawn::Text::Formatted::Arranger.new(self, options)
       arranger.consumed = Text::Formatted::Parser.format(string)
       arranger.finalize_line
