+++ lib/prawn/text/formatted.rb
@@ -1,4 +1,4 @@
-require_relative "../../pdf/core/text/formatted/wrap"
+require_relative "formatted/wrap"
 
 require_relative "formatted/box"
 require_relative "formatted/parser"
