+++ manual/document_and_page_options/page_size.rb
@@ -2,9 +2,9 @@
 #
 # Prawn comes with support for most of the common page sizes so you'll only need
 # to provide specific values if your intended format is not supported. To see a
-# list with all supported sizes take a look at: https://github.com/prawnpdf/prawn/blob/master/lib/prawn/document/page_geometry.rb
+# list with all supported sizes take a look at PDF::Core::PageGeometry
 #
-# To define the size use <code>:page_size</code> when creating new documents
+# # To define the size use <code>:page_size</code> when creating new documents
 # and <code>:size</code> when starting new pages. The default page size for new
 # documents is LETTER (612.00 x 792.00).
 #
